requirements.txt
streamlit
gspread
google-auth-oauthlib
Pillow
PyDrive2